F616D54A-BA0D-48E0-BAE2-269A4113D0AC
This PR removes any reliance on state in the backend of the
application itself. All information about the proof is encoded in a
string that is passed to the front end, when a command is to be
performed, that same string is passed back so that backend knows what
to operate on. This should make it much easier to have concurrent
users on the system. It's also just much cleaner