F999DB06-AA05-4D2C-BAE8-0BF885BC83F7
This PR updates the way that premise selection is handled for premise
operations. Before this PR all premise selection was in the view of
hte proof. Now, all premise operations open up a modal. This modal
allows for premise selection and in cases where necessary, the
specification of variables to instantiate.