D01BD36C-E447-4215-BB22-30C9857279D4
This PR adds support for `assert`. At any point in a proof a new
assertion can be made. This assertion is taken as a new goal. Once it
is proved the asserted formula can then be used as a premise.

Three tests have been added for this functionality. The update broke a
couple other tests which required updates.

This ended up also fixing some of the bugs that are in the todo file.