5012FF0F-92F4-4989-91F9-E7D53DA8B788 Add Tutorial
This commit adds a tutorial page to logos. The tutorial is designed
for someone who has some familiarity with philosophical logic to
acquaint themselves to the tool.

I couldn't help myself and added a more formal section on the
mathematical specification of formulas and the consequence relation.

It should be noted that the emacs plugin for this is now completely
broken but it could be fixed at some point.

Finally, I found a bug in disjunction-elimination that has been
fixed. The bug was a todo item, so that has also been closed.