F114BD39-B6D9-4E7E-9AA7-3513D0637583: Replace Dangerous Read Strings
This PR replaces any occurrence of read-string that occurrs in the
context of reading a formula with read-formula-string which is a safe
way of parsing a formula string. Additionally it removes all instances
of read-string from the premises file and replaces them with
Integer/parseInt. Finally any reading of an internal edn
representation of a structure has been replaced with edn/read-string.